scripts/requirements.txt
apache-airflow-providers-amazon>=8.0.0
boto3>=1.28.0